% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABA_script.R
\name{count_events}
\alias{count_events}
\title{Count the Number of Occurrences of Each Event in a Sequence}
\usage{
count_events(event_vector)
}
\arguments{
\item{event_vector}{A numeric vector representing a sequence of events.}
}
\value{
A dataframe with two columns: ID and Frequency, showing the number of occurrences of each event.
}
\description{
This function counts the number of occurrences of each unique event in a sequence.
The result is a dataframe with two columns: ID and Frequency.
}
\examples{
speaker_no <- c(3, 2, 3, 1, 4, 2, 4, 1, 4, 3, 2, 3)
count_events(speaker_no)

}
