% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABA_script.R
\name{create_is_aba}
\alias{create_is_aba}
\title{Create a Dummy Variable Indicating Whether a Code Represents 'ABA' (1) or not (0).}
\usage{
create_is_aba(codes_df)
}
\arguments{
\item{codes_df}{A dataframe of binary codes generated by the generate_codes function.}
}
\value{
A dataframe of codes with an additional column indicating whether the code represents 'ABA'.
}
\description{
This function creates a dummy variable indicating whether a code represents 'ABA'.
}
\examples{
speaker_no <- c(3, 2, 3, 1, 4, 2, 4, 1, 4, 3, 2, 3)
sequences_df <- generate_sequences(speaker_no, 3)
codes_df <- generate_codes(sequences_df)
create_is_aba(codes_df)

}
